NetworkConfig: Use QString as network id, not int
Yeah, once more no improvement but wasting resources.
Nah, looking better is an improvement too...and...

- Remove unneeded bool new_network - Great, is not it?
- Move needless public functions to private

+++ src/scanresults.cpp
@@ -277,7 +277,7 @@ void ScanResults::addNetwork() {
 	if (selectedNetworkId.isEmpty()) {
 		nc.newNetwork(selectedNetwork);
 	} else {
-		nc.editNetwork(selectedNetworkId.toInt(), selectedNetwork->text(1));
+		nc.editNetwork(selectedNetworkId, selectedNetwork->text(1));
 	}
 
 	nc.exec();
